+++ lib/core/database/database.dart
@@ -16,7 +16,7 @@ class Database {
         await Hive.initFlutter(_DATABASE_PATH);
         _registerAdapters();
         await _openBoxes();
-        if(profilesBox.keys.length == 0) setDefaults(clearAlerts: true);
+        if(profilesBox.keys.length == 0) setDefaults(clearEverything: true);
     }
 
     /// Deinitialize the database by closing all open hive boxes.
@@ -44,37 +44,44 @@ class Database {
 
     /// Open all Hive boxes for reading.
     Future<void> _openBoxes() async {
-        await Hive.openBox('lunasea');
         await Hive.openBox('alerts');
         await Hive.openBox<IndexerHiveObject>('indexers');
+        await Hive.openBox<LunaLogHiveObject>('logs');
+        await Hive.openBox('lunasea');
         await Hive.openBox<ProfileHiveObject>('profiles');
     }
 
     /// Set the default state for all hive boxes.
-    void setDefaults({ bool clearAlerts = false }) {
+    void setDefaults({ bool clearEverything = false }) {
         //Clear all the boxes
-        clearAllBoxes(clearAlerts: clearAlerts);
+        clearAllBoxes(clearEverything: clearEverything);
         //Set default profile & enabled profile
         profilesBox.put('default', ProfileHiveObject.empty());
         lunaSeaBox.put(LunaDatabaseValue.ENABLED_PROFILE.key, 'default');
     }
 
     //Get boxes
-    static Box get lunaSeaBox => Hive.box('lunasea');
     static Box get alertsBox => Hive.box('alerts');
-    static Box get profilesBox => Hive.box<ProfileHiveObject>('profiles');
-    static Box get indexersBox => Hive.box<IndexerHiveObject>('indexers');
+    static Box<IndexerHiveObject> get indexersBox => Hive.box<IndexerHiveObject>('indexers');
+    static Box<LunaLogHiveObject> get logsBox => Hive.box('logs');
+    static Box get lunaSeaBox => Hive.box('lunasea');
+    static Box<ProfileHiveObject> get profilesBox => Hive.box<ProfileHiveObject>('profiles');
 
     //Clear boxes
+    void clearAlertsBox() => alertsBox.deleteAll(alertsBox.keys);
+    void clearIndexersBox() => indexersBox.deleteAll(indexersBox.keys);
+    void clearLogsBox() => logsBox.deleteAll(logsBox.keys);
     void clearLunaSeaBox() => lunaSeaBox.deleteAll(lunaSeaBox.keys);
     void clearProfilesBox() => profilesBox.deleteAll(profilesBox.keys);
-    void clearIndexersBox() => indexersBox.deleteAll(indexersBox.keys);
-    void clearAlertsBox() => alertsBox.deleteAll(alertsBox.keys);
-    void clearAllBoxes({ bool clearAlerts = false }) {
+    
+    void clearAllBoxes({ bool clearEverything = false }) {
         clearLunaSeaBox();
         clearProfilesBox();
         clearIndexersBox();
-        if(clearAlerts) clearAlertsBox();
+        if(clearEverything) {
+            clearAlertsBox();
+            clearLogsBox();
+        }
     }
 
     //Profile values