+++ lib/core/extensions.dart
@@ -1,7 +1,6 @@
 export 'extensions/datetime.dart';
 export 'extensions/double.dart';
 export 'extensions/duration.dart';
-export 'extensions/flog_log_level.dart';
 export 'extensions/iap_product_details.dart';
 export 'extensions/int.dart';
 export 'extensions/navigator_state.dart';