+++ lib/core/models/configuration/indexer.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:hive/hive.dart';
 import 'package:lunasea/core.dart';
 
 part 'indexer.g.dart';