+++ lib/core/utilities.dart
@@ -2,5 +2,6 @@ export 'utilities/changelog.dart';
 export 'utilities/configuration.dart';
 export 'utilities/encryption.dart';
 export 'utilities/filesystem.dart';
+export 'utilities/logger.dart';
 export 'utilities/profile.dart';
 export 'utilities/uuid.dart';