+++ lib/modules/tautulli/routes/tautulli/route.dart
@@ -82,7 +82,7 @@ class _State extends State<_TautulliHomeRoute> {
     Widget get _appBar => LunaAppBar.dropdown(
         title: 'Tautulli',
         profiles: Database.profilesBox.keys.fold([], (value, element) {
-            if((Database.profilesBox.get(element) as ProfileHiveObject)?.tautulliEnabled ?? false) value.add(element);
+            if(Database.profilesBox.get(element)?.tautulliEnabled ?? false) value.add(element);
             return value;
         }),
         actions: context.read<TautulliState>().enabled ? [TautulliGlobalSettings()] : null,